some bugs + registrar faltante decimal
hacer migrate

+++ database/migrations/2019_08_04_154255_create_pedido_proveedor_clientes_table.php
@@ -28,7 +28,7 @@ class CreatePedidoProveedorClientesTable extends Migration
                     ->references('id')->on('pago_transportistas');
             $table->decimal('precio_galon_faltante', 9, 5)->nullable(); 
             $table->integer('asignacion');
-            $table->integer('faltante')->nullable();
+            $table->decimal('faltante')->nullable();
             $table->string('grifero')->nullable();
             $table->string('descripcion')->default('Faltante en grifo');
         });
